Split out `chat` cog's Config, Input, Output into separate files
